changed: `World.size` members can be floats

+++ src/two_d_library/world.py
@@ -26,7 +26,7 @@ class World:
     def __init__(
         self,
         *,
-        size_from_sequence: Sequence[int],
+        size_from_sequence: Sequence[float],
         centered_origin: bool = False,
     ) -> None:
         """Initialize a `World`."""
